requirements.txt
# Core
Flask==2.3.3
Werkzeug==2.3.8
python-dotenv==1.0.0

# Database
Flask-SQLAlchemy==3.0.4
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
SQLAlchemy==2.0.20

# Security
Flask-Talisman==1.0.0
Flask-Limiter==3.5.0
Flask-WTF==1.1.1
WTForms==3.0.1
email-validator==1.3.1
Flask-Login==0.6.3
itsdangerous==2.1.2
cryptography==39.0.1

# Utils
rapidfuzz==2.13.7
gunicorn==20.1.0
blinker==1.6.2

# Production
whitenoise==6.4.0
python-json-logger==2.0.7

# Security Headers
secure==1.3.0
